+++ src/components/Search.jsx
@@ -1,27 +1,8 @@
-import { useState } from 'react'
-
-const Search = () => {
-    const[query, setQuery] = useState('')
-
-    const searchResults = () => {
-        console.log("searched!")
-    }
-
-    const handleSubmit = e => {
-        e.preventDefault()
-        searchResults()
-    }
-    
-    const handleChange = e => {
-        setQuery(e.target.value)
-    }
-
-    console.log(query)
-
+const Search = (props) => {
     return (
         <div>
-            <form onSubmit={handleSubmit}>
-                <input type="text" value={query} onChange={handleChange}></input>
+            <form onSubmit={props.handleSubmit}>
+                <input type="text" value={props.query} onChange={props.handleChange}></input>
                 <input type="submit"></input>
             </form>
         </div>
